backend/requirements.txt
supabase
python-dotenv
requests
fastapi
uvicorn[standard]